Jsonifier Release v0.9.94
* Implemented a prettify function that utilizes simd instructions.
* Implemented a minify function that utilizes simd instructions.
* Implemented a validate function that utilizes simd instructions.
* Implemented a function for parsing into std::variant types.
* Implementing a fix to a buffer-overflow.

+++ Include/jsonifier/Vector.hpp
@@ -47,7 +47,7 @@ namespace jsonifier {
 
 		JSONIFIER_INLINE vector() {
 			if constexpr (doWeUseInitialBuffer) {
-				resize(static_cast<uint64_t>(static_cast<float>(1024 * 1024 * 4) * 0.85f));
+				resize(static_cast<uint64_t>(static_cast<double>(1024 * 1024 * 4) * 0.85f));
 			}
 		}
